ZenCart Sidebox to return to Mobile Site
Once a user is directed to the full/desktop site the cookie created needs
to be removed and the visitor redirected to see the mobile site.
This is particularly important because the origination cookie from the
mobile site does not have an expiration date/time, so may remain
indefinitely depending on the mobile device.

Change YOUR_TEMPLATE in each of the three folders under the includes folder
to reflect the template being used on your system then copy the includes
folder to the base of the ZenCart system.