Return to Mobile as default template
This is the remake of the ZenCart side of the Return To Mobile portion of code that provides a sidebox on the standard full site (once setup in the administration section.)  This will cause the sidebox to appear under all templates.  The filenames have been expanded to be more specific to minimize the likelihood that another program may overwrite them.