Update to README for default template style installation.
Updated the README.md file to reflect an installation that would put the files into the default like template not requiring any user interaction to install other than to make the sidebox visible from within admin utility.

Provided directions to also "properly" install the files, explaining that purpose is to make sure that the installation works.