Properly address the differences in catalogs for HTTPS and HTTP.
Not all websites use the same server base for both http and https, as a result the returned $_SESSION['REQUEST_URI'] may differ.  To address this and continue to verify against only the REQUEST_URI, the DIR_WS_HTTPS_CATALOG must also be used for comparison.

This is addressed by this Branch.  One improvement that may be made if considered acceptable is that the newly defined $Secure variable may be placed at the top of the file outside of each individual function.  Another programming "improvement" is that the $Secure variable should be deleted at the end of each procedure as it is equivalent to an Array after being set.