Want to compress mobile.php further
The idea of this Branch is to check to see if the file exists for the particular pagename, if it does not then will be able to say that a feature doesn't exist (rather than an undefined page)

If this can work, then would move all of the checks for pages such as checkout_success to the end of mobile.php because there are different filters for the various functions so have only made one function that uses the pagename.

There are still some issues with almost any of the pages having the anchor identifier in the URL which for users might cause questions to validity.  It would be great if this "concern" could be eliminated.